I did most of the CheckStyles for all my classes. Someneeds to do the mainapplication, homepanel, mainframe still. I am struggling with the catching and throwing exception. Maybe someone can help me. Please review this.
