+++ src/main/java/app/MainSearchApplication.java
@@ -2,7 +2,7 @@ package app;
 
 import data_access.DBSearchDataAccessObject;
 import data_access.PlayerStatisticsRepositoryImpl;
-import interface_adapter.PlayGameAspects.PlayGameController;
+import interface_adapter.play_game_aspects.PlayGameController;
 import interface_adapter.search.SearchViewModel;
 import use_case.playgame.FetchPlayerStatisticsInteractor;
 import use_case.playgame.FetchPlayerStatsUseCase;
