--- src/main/java/data_access/PlayerStatisticsAPI.java
@@ -1,35 +1,0 @@
-
-package data_access;
-
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.net.URLEncoder;
-
-public class PlayerStatisticsAPI {
-    private static final String BASE_URL = "https://samplejson.onrender.com/api/PlayerDataTotals/name/";
-
-    public String fetchPlayerData(String playerName) throws Exception {
-        String encodedName = URLEncoder.encode(playerName, "UTF-8").replace("+", "%20");
-        String urlString = BASE_URL + encodedName;
-
-        URL url = new URL(urlString);
-        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-        connection.setRequestMethod("GET");
-
-        int responseCode = connection.getResponseCode();
-        if (responseCode == HttpURLConnection.HTTP_OK) {
-            BufferedReader in = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-            StringBuilder response = new StringBuilder();
-            String inputLine;
-            while ((inputLine = in.readLine()) != null) {
-                response.append(inputLine);
-            }
-            in.close();
-            return response.toString();
-        } else {
-            throw new RuntimeException("GET request failed with response code: " + responseCode);
-        }
-    }
-}
