+++ src/main/java/entity/PlayerStatistic.java
@@ -1,5 +1,8 @@
 package entity;
 
+/**
+ * Represents the statistics of a player.
+ */
 public class PlayerStatistic {
     private final int age;
     private final int assists;
@@ -34,6 +37,42 @@ public class PlayerStatistic {
     private final int twoPointMade;
     private final double twoPointPercentage;
 
+    /**
+     * Constructs a new PlayerStatistic.
+     *
+     * @param age the age of the player
+     * @param assists the number of assists
+     * @param blocks the number of blocks
+     * @param defensiveRebounds the number of defensive rebounds
+     * @param effectiveFieldGoalPercentage the effective field goal percentage
+     * @param fieldGoalsAttempted the number of field goals attempted
+     * @param fieldGoalsMade the number of field goals made
+     * @param fieldGoalPercentage the field goal percentage
+     * @param freeThrowsMade the number of free throws made
+     * @param freeThrowsAttempted the number of free throws attempted
+     * @param freeThrowPercentage the free throw percentage
+     * @param gamesPlayed the number of games played
+     * @param gamesStarted the number of games started
+     * @param id the ID of the player
+     * @param minutesPlayed the number of minutes played
+     * @param offensiveRebounds the number of offensive rebounds
+     * @param personalFouls the number of personal fouls
+     * @param playerId the ID of the player
+     * @param playerName the name of the player
+     * @param points the number of points scored
+     * @param position the position of the player
+     * @param season the season year
+     * @param steals the number of steals
+     * @param team the team of the player
+     * @param threePointAttempts the number of three-point attempts
+     * @param threePointMade the number of three-point shots made
+     * @param threePointPercentage the three-point percentage
+     * @param totalRebounds the total number of rebounds
+     * @param turnovers the number of turnovers
+     * @param twoPointAttempts the number of two-point attempts
+     * @param twoPointMade the number of two-point shots made
+     * @param twoPointPercentage the two-point percentage
+     */
     public PlayerStatistic(int age, int assists, int blocks, int defensiveRebounds,
                            double effectiveFieldGoalPercentage, int fieldGoalsAttempted,
                            int fieldGoalsMade, double fieldGoalPercentage, int freeThrowsMade,
@@ -207,47 +246,54 @@ public class PlayerStatistic {
         return twoPointPercentage;
     }
 
+    /**
+     * Formats the player statistics into a readable format.
+     *
+     * @return a string containing the formatted player statistics
+     */
     public String formatPlayerStats() {
-        StringBuilder formattedStats = new StringBuilder();
+        final StringBuilder formattedStats = new StringBuilder();
+        final String nl = "\n";
+        final String dl = "\n\n";
 
-        formattedStats.append("Player Name: ").append(getPlayerName()).append("\n");
-        formattedStats.append("Age: ").append(getAge()).append("\n");
-        formattedStats.append("Position: ").append(getPosition()).append("\n");
-        formattedStats.append("Team: ").append(getTeam()).append("\n");
-        formattedStats.append("Season: ").append(getSeason()).append("\n\n");
+        formattedStats.append("Player Name: ").append(getPlayerName()).append(nl);
+        formattedStats.append("Age: ").append(getAge()).append(nl);
+        formattedStats.append("Position: ").append(getPosition()).append(nl);
+        formattedStats.append("Team: ").append(getTeam()).append(nl);
+        formattedStats.append("Season: ").append(getSeason()).append(dl);
 
-        formattedStats.append("Games Played: ").append(getGamesPlayed()).append("\n");
-        formattedStats.append("Games Started: ").append(getGamesStarted()).append("\n");
-        formattedStats.append("Minutes Played: ").append(getMinutesPlayed()).append("\n\n");
+        formattedStats.append("Games Played: ").append(getGamesPlayed()).append(nl);
+        formattedStats.append("Games Started: ").append(getGamesStarted()).append(nl);
+        formattedStats.append("Minutes Played: ").append(getMinutesPlayed()).append(dl);
 
-        formattedStats.append("Field Goals Made: ").append(getFieldGoalsMade()).append("\n");
-        formattedStats.append("Field Goal Attempts: ").append(getFieldGoalsAttempted()).append("\n");
-        formattedStats.append("Field Goal Percentage: ").append(getFieldGoalPercentage()).append("\n\n");
+        formattedStats.append("Field Goals Made: ").append(getFieldGoalsMade()).append(nl);
+        formattedStats.append("Field Goal Attempts: ").append(getFieldGoalsAttempted()).append(nl);
+        formattedStats.append("Field Goal Percentage: ").append(getFieldGoalPercentage()).append(dl);
 
-        formattedStats.append("Three-Point FG Made: ").append(getThreePointMade()).append("\n");
-        formattedStats.append("Three-Point Attempts: ").append(getThreePointAttempts()).append("\n");
-        formattedStats.append("Three-Point Percentage: ").append(getThreePointPercentage()).append("\n\n");
+        formattedStats.append("Three-Point FG Made: ").append(getThreePointMade()).append(nl);
+        formattedStats.append("Three-Point Attempts: ").append(getThreePointAttempts()).append(nl);
+        formattedStats.append("Three-Point Percentage: ").append(getThreePointPercentage()).append(dl);
 
-        formattedStats.append("Two-Point FG Made: ").append(getTwoPointMade()).append("\n");
-        formattedStats.append("Two-Point Attempts: ").append(getTwoPointAttempts()).append("\n");
-        formattedStats.append("Two-Point Percentage: ").append(getTwoPointPercentage()).append("\n");
-        formattedStats.append("Effective Field Goal Percentage: ").append(getEffectiveFieldGoalPercentage()).append("\n\n");
+        formattedStats.append("Two-Point FG Made: ").append(getTwoPointMade()).append(nl);
+        formattedStats.append("Two-Point Attempts: ").append(getTwoPointAttempts()).append(nl);
+        formattedStats.append("Two-Point Percentage: ").append(getTwoPointPercentage()).append(nl);
+        formattedStats.append("Effective Field Goal Percentage: ").append(getEffectiveFieldGoalPercentage()).append(dl);
 
-        formattedStats.append("Free Throws Made: ").append(getFreeThrowsMade()).append("\n");
-        formattedStats.append("Free Throw Attempts: ").append(getFreeThrowsAttempted()).append("\n");
-        formattedStats.append("Free Throw Percentage: ").append(getFreeThrowPercentage()).append("\n\n");
+        formattedStats.append("Free Throws Made: ").append(getFreeThrowsMade()).append(nl);
+        formattedStats.append("Free Throw Attempts: ").append(getFreeThrowsAttempted()).append(nl);
+        formattedStats.append("Free Throw Percentage: ").append(getFreeThrowPercentage()).append(dl);
 
-        formattedStats.append("Offensive Rebounds: ").append(getOffensiveRebounds()).append("\n");
-        formattedStats.append("Defensive Rebounds: ").append(getDefensiveRebounds()).append("\n");
-        formattedStats.append("Total Rebounds: ").append(getTotalRebounds()).append("\n\n");
+        formattedStats.append("Offensive Rebounds: ").append(getOffensiveRebounds()).append(nl);
+        formattedStats.append("Defensive Rebounds: ").append(getDefensiveRebounds()).append(nl);
+        formattedStats.append("Total Rebounds: ").append(getTotalRebounds()).append(dl);
 
-        formattedStats.append("Assists: ").append(getAssists()).append("\n");
-        formattedStats.append("Steals: ").append(getSteals()).append("\n");
-        formattedStats.append("Blocks: ").append(getBlocks()).append("\n\n");
+        formattedStats.append("Assists: ").append(getAssists()).append(nl);
+        formattedStats.append("Steals: ").append(getSteals()).append(nl);
+        formattedStats.append("Blocks: ").append(getBlocks()).append(dl);
 
-        formattedStats.append("Turnovers: ").append(getTurnovers()).append("\n");
-        formattedStats.append("Personal Fouls: ").append(getPersonalFouls()).append("\n");
-        formattedStats.append("Points Scored: ").append(getPoints()).append("\n");
+        formattedStats.append("Turnovers: ").append(getTurnovers()).append(nl);
+        formattedStats.append("Personal Fouls: ").append(getPersonalFouls()).append(nl);
+        formattedStats.append("Points Scored: ").append(getPoints()).append(nl);
 
         return formattedStats.toString();
     }
