+++ src/main/java/use_case/playgame/FetchPlayerStatisticsInteractor.java
@@ -1,11 +1,15 @@
 package use_case.playgame;
 
-import entity.PlayerStatistic;
-import interface_adapter.PlayGameAspects.PlayerStatisticsRepository;
-
 import java.util.List;
 import java.util.stream.Collectors;
 
+import entity.PlayerStatistic;
+import interface_adapter.play_game_aspects.PlayerStatisticsRepository;
+
+/**
+ * Interactor for fetching player statistics.
+ * Implements the input boundary interface to handle the use case logic.
+ */
 public class FetchPlayerStatisticsInteractor implements FetchPlayerStatisticsInputBoundary {
     private final PlayerStatisticsRepository repository;
 
@@ -25,13 +29,13 @@ public class FetchPlayerStatisticsInteractor implements FetchPlayerStatisticsInp
 
     @Override
     public List<String> getAvailableYears(String playerName) {
-        List<Integer> years = repository.fetchAvailableYearsForPlayer(playerName);
+        final List<Integer> years = repository.fetchAvailableYearsForPlayer(playerName);
         return years.stream().map(String::valueOf).collect(Collectors.toList());
     }
 
     @Override
     public double getAverageStat(String playerName, String year, String statType) {
-        PlayerStatistic stat = repository.fetchStatsForPlayerByYear(playerName, Integer.parseInt(year));
+        final PlayerStatistic stat = repository.fetchStatsForPlayerByYear(playerName, Integer.parseInt(year));
         if (stat == null) {
             throw new IllegalArgumentException("No data available for year " + year);
         }
@@ -40,15 +44,20 @@ public class FetchPlayerStatisticsInteractor implements FetchPlayerStatisticsInp
     }
 
     private double calculateAverage(PlayerStatistic stat, String statType) {
+        final double result;
         switch (statType) {
             case "Average Rebounds":
-                return (double) stat.getTotalRebounds() / stat.getGamesPlayed();
+                result = (double) stat.getTotalRebounds() / stat.getGamesPlayed();
+                break;
             case "Average Points":
-                return (double) stat.getPoints() / stat.getGamesPlayed();
+                result = (double) stat.getPoints() / stat.getGamesPlayed();
+                break;
             case "Average Assists":
-                return (double) stat.getAssists() / stat.getGamesPlayed();
+                result = (double) stat.getAssists() / stat.getGamesPlayed();
+                break;
             default:
                 throw new IllegalArgumentException("Invalid statistic type.");
         }
+        return result;
     }
 }
