+++ src/main/java/use_case/playgame/FetchPlayerStatsUseCase.java
@@ -2,6 +2,10 @@ package use_case.playgame;
 
 import entity.PlayerStatistic;
 
+/**
+ * Use case for fetching player statistics by year.
+ * Acts as an intermediary between the controller and the interactor.
+ */
 public class FetchPlayerStatsUseCase {
     private final FetchPlayerStatisticsInputBoundary interactor;
 
@@ -9,6 +13,13 @@ public class FetchPlayerStatsUseCase {
         this.interactor = interactor;
     }
 
+    /**
+     * Executes the use case to fetch player statistics for the specified year.
+     *
+     * @param playerName the name of the player
+     * @param year the year for which to fetch statistics
+     * @return the player's statistics for the specified year
+     */
     public PlayerStatistic execute(String playerName, String year) {
         return interactor.fetchPlayerStatisticsByYear(playerName, Integer.parseInt(year));
     }
